Add files via upload
This project is a QR Code Generator that allows users to create QR codes based on input text. The website provides additional functionality to resize the QR code and download it for personal use. It's designed to be user-friendly and efficient, making it a handy tool for anyone who needs to generate QR codes on the fly.

->Features

1.Generate QR Codes: Enter any text or URL, and the tool will generate a corresponding QR code.
2.Adjustable Size: Resize the QR code based on your needs by choosing custom dimensions.
3.Download QR Code: Easily download the generated QR code as an image file.
4.Responsive Design: Works smoothly across desktop and mobile devices.

->Technologies Used

HTML/CSS/JavaScript: Frontend structure and styling.